requirements.txt
fastapi
uvicorn
gitpython
psutil
